Create code missing contacts
This program i believe is calling upon code when you enter 2 that either does not exist or is not saved after entry the program i think needs to contain a save feature which i may include upon a later date but here it is in it's current state. Also i do not know if the arrays were declared anywhere in here.